requirements.txt
# Leaf Health Check - Production Dependencies

# Frontend & Web Framework
streamlit==1.28.1
streamlit-option-menu==0.3.2

# Deep Learning & AI
tensorflow==2.13.1
keras==2.13.1
numpy==1.24.3
scikit-learn==1.3.0

# Image Processing
opencv-python==4.8.0.76
Pillow==10.0.0

# Database
sqlite3  # Built-in with Python

# Data Processing
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7

# Development & Testing (optional, comment out for production)
pytest==7.4.0
black==23.7.0
pylint==2.17.4
flake8==6.0.0

# Deployment
gunicorn==21.2.0
python-multipart==0.0.6

# Google Gemini AI Integration
google-generativeai==0.3.0

# Optional: GPU support (uncomment if using CUDA)
# tensorflow-gpu==2.13.1
